Travis: build and test against Node 6+ only
The travis versions are the supported -development environment-
versions. Since this is a front-end-only component, it does not have any
implications for versions of node used in production.

This change means we can use async/await compiled to native generators
in tests. When Node gets native async/await, we will support that
version and newer only for development.